requirements.txt
# Core Python Dependencies
python-dotenv==1.0.0

# LangChain Integration (wrappers only - no behavioral changes)
langchain==0.1.0
langchain-community==0.0.13
langgraph==0.0.20

# Vector Database and Embeddings
faiss-cpu==1.9.0
sentence-transformers==3.0.1
transformers==4.36.2
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# Document Processing
PyPDF2==3.0.1
pypdf==3.17.4

# Text Processing
rank-bm25==0.2.2

# Data Validation
pydantic==2.5.3

# Evaluation Metrics
scikit-learn==1.3.2
rouge-score==0.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Data Handling
pandas==2.1.4
numpy==1.26.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# API (Optional)
fastapi==0.108.0
uvicorn==0.25.0

# Utilities
tqdm==4.66.1
requests==2.31.0
pyyaml==6.0.1

# Logging
colorlog==6.8.0

# Note: Ollama must be installed separately
# Download from: https://ollama.ai/download
# Then run: ollama pull llama3.1:8b